requirements.txt
seaborn
pandas
numpy 
scikit-learn==1.4.0
scipy==1.11.1
pillow==10.0.0
matplotlib==3.7.1
ray[serve]==2.9.2
kserve==0.12.0